+++ src/cart/src/main/java/works/weave/socks/cart/repositories/ItemEntity.java
@@ -5,5 +5,5 @@ public interface ItemEntity {
 
     int getQuantity();
 
-    float getUnitPrice();
+    int getUnitPrice();
 }
